requirements.txt
aiodns==2.0.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
discord.py==1.7.3
idna==3.1
multidict==5.1.0
pip==21.2.2
pycares==4.0.0
pycparser==2.20
setuptools==58.0.4
slackclient==2.9.3
typing-extensions==3.10.0.2
wheel==0.37.0
yarl==1.6.3
